Echo Bot/requirements.txt
gunicorn==20.0.4
Flask==1.1.1
line-bot-sdk==1.16.0